HTML の表示・非表示の制御部分をコンポーネントに切り分けた
